fixup __main__.py
rename __main__.py to main.py
rename main.py to stats.py
